feat(projects): injeta o prisma no projectService para fazer consultas ao banco de dados e adiciona validação de id nos controllers

+++ src/modules/projects/projects.controller.ts
@@ -4,6 +4,7 @@ import {
   Delete,
   Get,
   HttpCode,
+  HttpException,
   HttpStatus,
   Param,
   ParseUUIDPipe,
@@ -32,8 +33,12 @@ export class ProjectsController {
   @ApiResponse({
     type: ProjectListItemDto,
   })
-  findOne(@Param('id', ParseUUIDPipe) id: string) {
-    return this.projectsService.findById(id);
+  async findOne(@Param('id', ParseUUIDPipe) id: string) {
+    const project = await this.projectsService.findById(id);
+    if (!project) {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND);
+    }
+    return project;
   }
 
   @Post()
@@ -48,13 +53,24 @@ export class ProjectsController {
   @ApiResponse({
     type: ProjectListItemDto,
   })
-  update(@Param('id', ParseUUIDPipe) id: string, @Body() data: ProjectRequestDto) {
+  async update(@Param('id', ParseUUIDPipe) id: string, @Body() data: ProjectRequestDto) {
+    const project = await this.projectsService.findById(id);
+
+    if (!project) {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND);
+    }
+
     return this.projectsService.update(id, data);
   }
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  remove(@Param('id', ParseUUIDPipe) id: string) {
+  async remove(@Param('id', ParseUUIDPipe) id: string) {
+    const project = await this.projectsService.findById(id);
+    if (!project) {
+      throw new HttpException('Project not found', HttpStatus.NOT_FOUND);
+    }
+
     return this.projectsService.remove(id);
   }
 }
